cpm/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "context-packet-manager"
version = "0.0.1"
description = "Context Packet Manager - Modular Context packets for LLM retrievve"
readme = "README.md"
requires-python = ">=3.10,<3.13"
authors = [{ name = "Andrea Redegalli" }]
keywords = ["context-component", "rag", "context-manager", "semantic-search", "context-vector", "code-analysis", "python"]

dependencies = [
    "fastapi>=0.110,<1",
    "uvicorn[standard]>=0.27,<1",
    "pydantic>=2.6,<3",
    "pyyaml>=6,<7",
    "requests>=2.31,<3",
    "numpy>=1.26,<3",
    "faiss-cpu>=1.7,<2",
    "transformers>=4.41,<5",
    "tokenizers>=0.15,<1",
    "huggingface-hub>=0.20,<1",
    "safetensors>=0.4,<1",
    "accelerate>=0.26,<2",
    "sentence-transformers>=2.3,<4",
    "tree-sitter>=0.20,<1",
    "tree-sitter-language-pack>=0.0.6,<1",
    "mistune>=3.0,<4",
    "packaging>=23,<25",
    "rich>=13,<15",
    "structlog>=24,<25",
]

[project.optional-dependencies]
cpu = ["torch>=2.1,<3"]
gpu = ["torch>=2.1,<3"]
faiss-gpu = ["faiss-gpu>=1.7,<2"]
treesitter-alt = ["tree-sitter-languages>=1.9,<2"]

dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=4,<6",
    "ruff>=0.4,<1",
    "mypy>=1.8,<2",
    "black>=24,<25",
    "httpx>=0.26,<1",
    "types-PyYAML",
    "types-requests",
]

docs = ["mkdocs>=1.5,<2", "mkdocs-material>=9,<10", "mkdocstrings[python]>=0.24,<1"]

all = [
    "torch>=2.1,<3",
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=4,<6",
    "ruff>=0.4,<1",
    "mypy>=1.8,<2",
    "black>=24,<25",
    "httpx>=0.26,<1",
    "types-PyYAML",
    "types-requests",
    "mkdocs>=1.5,<2",
    "mkdocs-material>=9,<10",
    "mkdocstrings[python]>=0.24,<1",
]

[project.scripts]
cpm = "cpm_cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
rag = ["py.typed"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008", "B905"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["B011"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["tree_sitter.*", "tree_sitter_language_pack.*", "tree_sitter_languages.*", "mistune.*", "faiss.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = ["--strict-markers", "--strict-config", "--cov=rag", "--cov-report=term-missing", "--cov-report=html"]
markers = ["slow", "integration", "requires_gpu"]

[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
